+++ src/main/java/com/gitlab/cdagaming/craftpresence/handler/curse/ManifestHandler.java
@@ -1,6 +1,7 @@
 package com.gitlab.cdagaming.craftpresence.handler.curse;
 
 import com.gitlab.cdagaming.craftpresence.Constants;
+import com.gitlab.cdagaming.craftpresence.CraftPresence;
 import com.gitlab.cdagaming.craftpresence.handler.FileHandler;
 import com.gitlab.cdagaming.craftpresence.handler.StringHandler;
 import net.minecraft.client.resources.I18n;
@@ -27,6 +28,7 @@ public class ManifestHandler {
 
             if (manifest != null && !StringHandler.isNullOrEmpty(manifest.name)) {
                 manifests.put(manifest.name, manifest);
+                CraftPresence.packFound = true;
                 Constants.LOG.info(I18n.format("craftpresence.logger.info.manifest.loaded", manifest.name));
             }
         } catch (Exception ex) {